Process:getPid() should return null for empty process status
This makes the return value consistent with the method documentation. Note that we don't check that the 'pid' property exists for the same reason we don't check other fields in updateStatus(). By that point, we should assume proc_get_status()'s return value is valid and includes the array keys mentioned in its documentation.