Backwards compatibility with Promise v1 API
This simple patch updates this project to target either the current
Promise v2 API or the legacy Promise v1 API. While we actively encourage
people to use the current Promise v2 API, we still provide active
support for the old Promsie v1 API. For the most part, they are fully
compatible, see also the
https://github.com/reactphp/promise/blob/2.x/CHANGELOG.md for details.

The legacy Promise v1 API also supports outdated legacy PHP 5.3 which we
understand is still a requirement for a certain non-zero number of
projects. That being said, we actively encourage our users to update to
the latest version of PHP 7+ for best performance.