Skip affected test cases to avoid bogus test failure on PHP 7.4
The build error on PHP 7.4 will be addressed via
https://github.com/reactphp/dns/pull/164. Interestingly, it fails in the
`FunctionalResolverTest` class and does not fail anymore by excluding
some other test cases. It can only be reproduced by running:

```
$ vendor/bin/phpunit --coverage-text --exclude-group internet
```

It can not be reproduced when running on any other PHP version, when
`ext-xdebug` is disabled, when coverage is not collected or when not
excluding this group.

This means the test failure is bogus.

We work around this by temporarily skipping some test case for now.